schedulebot: Move generic zulip api code into its own file
for re-use
